removed more code duplications
