+++ src/Core/Types.Filters/Expressions/String/StringStartsWithOperationHandler.cs
@@ -1,48 +1,34 @@
 using System.Linq.Expressions;
-using HotChocolate.Language;
-using HotChocolate.Utilities;
 
 namespace HotChocolate.Types.Filters.Expressions
 {
-    public class StringStartsWithOperationHandler
-        : IExpressionOperationHandler
+    public sealed class StringStartsWithOperationHandler
+        : StringOperationHandlerBase
     {
-
-        public bool TryHandle(
+        protected override bool TryCreateExpression(
             FilterOperation operation,
-            IInputType type,
-            IValueNode value,
-            Expression instance,
-            ITypeConversion converter,
-            out Expression expression)
+            MemberExpression property,
+            object parsedValue, out
+            Expression expression)
         {
-            if (operation.Type == typeof(string)
-                && (value is StringValueNode || value.IsNull()))
+            switch (operation.Kind)
             {
-                object parsedValue = type.ParseLiteral(value);
+                case FilterOperationKind.StartsWith:
+                    expression = FilterExpressionBuilder.StartsWith(
+                        property, parsedValue);
+                    return true;
 
-                MemberExpression property =
-                    Expression.Property(instance, operation.Property);
+                case FilterOperationKind.NotStartsWith:
+                    expression = FilterExpressionBuilder.Not(
+                        FilterExpressionBuilder.StartsWith(
+                            property, parsedValue)
+                    );
+                    return true;
 
-                switch (operation.Kind)
-                {
-
-                    case FilterOperationKind.StartsWith:
-                        expression = FilterExpressionBuilder.StartsWith(
-                            property, parsedValue);
-                        return true;
-
-                    case FilterOperationKind.NotStartsWith:
-                        expression = FilterExpressionBuilder.Not(
-                            FilterExpressionBuilder.StartsWith(
-                                property, parsedValue)
-                        );
-                        return true;
-                }
+                default:
+                    expression = null;
+                    return false;
             }
-
-            expression = null;
-            return false;
         }
     }
 }
